+++ src/modules/Builder/features/ActivitySettings/ActivitySettings.utils.tsx
@@ -2,16 +2,12 @@ import { lazy } from 'react';
 
 import { Svg } from 'shared/components/Svg';
 import { SettingParam } from 'shared/utils';
-import { ItemFormValues } from 'modules/Builder/types';
-import { ItemResponseType } from 'shared/consts';
-import { Item, SingleSelectionConfig, MultipleSelectionConfig, SliderConfig } from 'shared/state';
 import { Item as ItemNavigation } from 'shared/components/NavigationMenu/NavigationMenu.types';
 import { Mixpanel } from 'shared/utils/mixpanel';
 
 import { SubscalesConfiguration } from './SubscalesConfiguration';
 import { ScoresAndReports } from './ScoresAndReports';
 import { GetActivitySettings } from './ActivitySettings.types';
-import { ItemsWithScore } from './ScoresAndReports/ScoreContent/ScoreContent.types';
 
 const ReportConfigSetting = lazy(() => import('modules/Builder/features/ReportConfigSetting'));
 
@@ -65,11 +61,3 @@ export const getActivitySettings = ({
     },
   ];
 };
-
-export const checkOnItemTypeAndScore = (item: ItemFormValues | Item): item is ItemsWithScore =>
-  (item.config as SingleSelectionConfig | MultipleSelectionConfig | SliderConfig)?.addScores &&
-  [
-    ItemResponseType.SingleSelection,
-    ItemResponseType.MultipleSelection,
-    ItemResponseType.Slider,
-  ].includes(item.responseType);
