+++ src/modules/Dashboard/features/Applet/Schedule/Legend/Legend.utils.test.tsx
@@ -1,6 +1,6 @@
 import { screen } from '@testing-library/react';
 
-import { renderWithProviders } from 'shared/utils';
+import { renderWithProviders } from 'shared/utils/renderWithProviders';
 
 import { getTitle } from './Legend.utils';
