+++ src/shared/features/AppletSettings/ExportDataSetting/Popups/ExportSettingsPopup/ExportSettingsPopup.tsx
@@ -6,7 +6,8 @@ import { useFormContext } from 'react-hook-form';
 import { Svg } from 'shared/components/Svg';
 import { applet } from 'shared/state';
 import { SelectController } from 'shared/components/FormComponents';
-import { DatePicker, Modal } from 'shared/components';
+import { DatePicker } from 'shared/components/DatePicker';
+import { Modal } from 'shared/components/Modal';
 import { theme, StyledBodyLarge, StyledFlexTopCenter, StyledModalWrapper } from 'shared/styles';
 import { SelectEvent } from 'shared/types';
 import { DateType } from 'shared/components/DatePicker/DatePicker.types';
